requirements.txt
appdirs==1.4.3
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.11
django-braces==1.11.0
django-extensions==1.7.9
django-oauth-toolkit==1.0.0
django-rest-framework-social-oauth2==1.0.8
django-webpack-loader==0.4.1
djangorestframework==3.6.2
gunicorn==19.7.1
httpie==0.9.9
oauthlib==2.0.2
olefile==0.44
packaging==16.8
Pillow==4.1.1
Pygments==2.2.0
PyJWT==1.5.2
pyparsing==2.2.0
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2017.2
requests==2.14.2
requests-oauthlib==0.8.0
six==1.10.0
social-auth-app-django==1.2.0
social-auth-core==1.4.0
whitenoise==3.3.0
